Fix VTK files for direction dependent degree cells
To be able to render direction dependent degree cells Paraview needs to
know what array contains the high order degrees.  This is set by adding
an attribute to the cell data tag.  Using WritVTK this is done with

```julia
vtk[VTKCellData()] = Dict("HigherOrderDegrees" => "HigherOrderDegrees")
```

This allows the meshes to be viewed in current versions of Paraview
(5.10 and 5.11).